refactor(runtimed): split peer runtime agent handler
