+++ crates/runtimed/src/notebook_sync_server/mod.rs
@@ -57,6 +57,7 @@ mod peer;
 mod peer_notebook_sync;
 mod peer_pool_sync;
 mod peer_presence;
+mod peer_runtime_agent;
 mod peer_runtime_sync;
 mod peer_session;
 mod peer_writer;
@@ -75,6 +76,7 @@ pub use path_index::{PathIndex, PathIndexError};
 pub(crate) use peer::*;
 #[cfg(test)]
 pub(crate) use peer_presence::sanitize_peer_label;
+pub(crate) use peer_runtime_agent::handle_runtime_agent_sync_connection;
 pub(crate) use peer_runtime_sync::notebook_execution_context_id;
 pub(crate) use persist::*;
 pub(crate) use project_context::refresh_project_context_on_save_as;
